chore(cluster-init): refactor prowplugin code
