+++ cmd/cluster-init/cmd/onboard/prowplugin/update_test.go
@@ -1,4 +1,4 @@
-package onboard
+package prowplugin
 
 import (
 	"testing"
